+++ src/Perch.App/PlatformServices.cs
@@ -17,6 +17,7 @@ namespace Perch.Avalonia;
 internal static class PlatformServices
 {
     public static IWindowActivator WindowActivator { get; } = new Impl.WindowActivator();
+    public static ISessionLauncher SessionLauncher { get; } = new Impl.SessionLauncher();
     public static IPathInstaller PathInstaller { get; } = new Impl.PathInstaller();
     public static IAudioCue AudioCue { get; } = new Impl.AudioCue();
     public static IWindowChrome WindowChrome { get; } = new Impl.WindowChrome();